new page product-description

+++ src/api/product.api.js
@@ -1,7 +1,20 @@
 import { API_URL } from "../config/env";
+import { getAuthToken } from "@/lib/helpers";
+
+
+// Buscar productos por término
+export const searchProducts = async (term) => {
+  const response = await fetch(`${API_URL}/api/product/search?term=${encodeURIComponent(term)}`, {
+    method: "GET",
+  });
+
+  if (!response.ok) throw new Error("Error al buscar productos");
+
+  return response.json();
+};
 
 export const getProductById = async (id) => {
-  const response = await fetch(`${API_URL}/api/products/${id}`, {
+  const response = await fetch(`${API_URL}/api/product/${id}`, {
     method: "GET",
   });
 
@@ -13,7 +26,7 @@ export const getProductById = async (id) => {
 };
 
 export const getAllProducts = async () => {
-  const response = await fetch(`${API_URL}/api/products`, {
+  const response = await fetch(`${API_URL}/api/product`, {
     method: "GET",
   });
 
@@ -25,7 +38,10 @@ export const getAllProducts = async () => {
 };
 
 export const createProduct = async (product) => {
-  const response = await fetch(`${API_URL}/api/products`, {
+  const token = getAuthToken();
+  if (!token) throw new Error("Token no disponible");
+
+  const response = await fetch(`${API_URL}/api/product/create`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -42,7 +58,10 @@ export const createProduct = async (product) => {
 };
 
 export const updateProduct = async (id, product) => {
-  const response = await fetch(`${API_URL}/api/products/${id}`, {
+  const token = getAuthToken();
+  if (!token) throw new Error("Token no disponible");
+
+  const response = await fetch(`${API_URL}/api/product/${id}`, {
     method: "PUT",
     headers: {
       "Content-Type": "application/json",
@@ -59,7 +78,10 @@ export const updateProduct = async (id, product) => {
 };
 
 export const removeProduct = async (id) => {
-  const response = await fetch(`${API_URL}/api/products/${id}`, {
+  const token = getAuthToken();
+  if (!token) throw new Error("Token no disponible");
+
+  const response = await fetch(`${API_URL}/api/product/${id}`, {
     method: "DELETE",
     headers: {
       "Content-Type": "application/json",
